# public/openapi.yaml
openapi: 3.0.3
info:
  title: allaboard.dev API
  version: "1.0"
  description: REST API for the Allaboard board-climbing community platform.
servers:
  - url: https://www.allaboard.dev
    description: Production (www)
  - url: https://allaboard.dev
    description: Production (apex)
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: allaboard_session
      description: iron-session encrypted cookie set after Google OAuth login.
    apiKeyAuth:
      type: apiKey
      in: query
      name: token
      description: Personal API token from your profile page. Pass as ?token=<value>.
paths:
...
  /api/feed:
    get:
      summary: Fetch the global activity feed (most recent 50 ticks, newest first).
      responses:
        "200":
          description: "Array of feed activity objects: ``json [ { \"id\": \"uuid\", \"date\": \"2026-04-01T...\", \"sent\": true, \"rating\": 3, \"comment\": \"Great problem!\", \"suggestedGrade\": \"V5\", \"instagramUrl\": \"https://www.instagram.com/reel/...\", \"attempts\": 4, \"user\": { \"id\": \"alex\", \"handle\": \"alex\", \"displayName\": \"Alex\", ... }, \"climb\": { \"id\": \"uuid\", \"name\": \"The Riddler\", \"grade\": \"V5\", ... } } ] ``"
        "500":
          description: database error.
...
  /api/inbox/{id}:
    patch:
      summary: Mark an inbox item as read.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "{ read: true } on success."
        "401":
          description: not authenticated.
        "403":
          description: not the inbox owner.
        "404":
          description: item does not exist.
        "500":
          description: database error.
      security:
        - apiKeyAuth: []
        - cookieAuth: []
...